Fixes for builds on Windows-2022
Add explicit typecasts in support for 32-bit architectures

Winsock compatibility in EVPath

+++ source/adios2/toolkit/format/bp5/BP5Serializer.cpp
@@ -1629,16 +1629,16 @@ std::vector<core::iovec> BP5Serializer::BreakoutContiguousMetadata(
         {
             uint64_t MEBSize;
             helper::CopyFromBuffer(Aggregate.data(), Position, &MEBSize);
-            MetadataBlocks.push_back({Aggregate.data() + Position, MEBSize});
-            Position += MEBSize;
+            MetadataBlocks.push_back({Aggregate.data() + Position, static_cast<size_t>(MEBSize)});
+            Position += static_cast<size_t>(MEBSize);
         }
         helper::CopyFromBuffer(Aggregate.data(), Position, &ABCount);
         for (uint64_t i = 0; i < ABCount; ++i)
         {
             uint64_t AEBSize;
             helper::CopyFromBuffer(Aggregate.data(), Position, &AEBSize);
-            AttributeBlocks.push_back({Aggregate.data() + Position, AEBSize});
-            Position += AEBSize;
+            AttributeBlocks.push_back({Aggregate.data() + Position, static_cast<size_t>(AEBSize)});
+            Position += static_cast<size_t>(AEBSize);
         }
         uint64_t element;
         helper::CopyFromBuffer(Aggregate.data(), Position, &DSCount);
